Material Canvas: Moving graph view class to Atom Tools Framework Migrated material canvas graph view to atom tools framework to make it available for the pass editor and other tools. It’s an amalgamation of features from graph model, graph canvas, and script canvas main window classes that only focuses on one graph at a time to be compatible with the document/view system.
