--- Gems/Atom/Tools/MaterialCanvas/Code/Source/MaterialCanvasUtil.h
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) Contributors to the Open 3D Engine Project.
- * For complete copyright and license terms please see the LICENSE at the root of this distribution.
- *
- * SPDX-License-Identifier: Apache-2.0 OR MIT
- *
- */
-
-#pragma once
-
-#if !defined(Q_MOC_RUN)
-#include <GraphCanvas/Widgets/NodePalette/TreeItems/NodePaletteTreeItem.h>
-#endif
-
-namespace MaterialCanvas
-{
-    using CreateNodePaletteItemsCallback = AZStd::function<GraphCanvas::GraphCanvasTreeItem*(const AZ::Crc32&)>;
-} // namespace MaterialCanvas
